+++ src/main/java/com/auction/shared/model/Electronics.java
@@ -1,5 +1,6 @@
 package com.auction.shared.model;
 
+import java.math.BigDecimal;
 import java.util.LinkedHashMap;
 import java.util.Map;
 
@@ -8,7 +9,7 @@ public class Electronics extends Item {
     private String brand;
     private String condition;
 
-    public Electronics(String title, String description, double startingPrice, String brand, String condition) {
+    public Electronics(String title, String description, BigDecimal startingPrice, String brand, String condition) {
         super(title, description, startingPrice, ItemCategory.ELECTRONICS);
         this.brand = brand;
         this.condition = condition;
